增加获得选手信息

+++ 02-javaee/javaee-04-springboot/springboot-01-jpa/src/main/java/cn/gaohank/program/dao/PlayersDao.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PlayersDao extends JpaRepository<Players, Long> {
+public interface PlayersDao extends JpaRepository<Players, Integer> {
 }
